relocatable toolchain
create a toolchain using --with-sysroot.
the resulting toolchain can be used standalone (i.e. as a native
static i386 compiler on a x64 system) or put in another directory,
as no paths are hardcoded.
the toolchain is located in /opt/toolchain.
the pkg/relocate_toolchain will link all files in the toolchain dir
into their respective / locations.
a slight disadvantage of that approach is that gcc now searches first
in its own toolchain path for headers, and the same for ld and libs.
therefore we use the musl-gcc wrapper as official "gcc" link, while
the real GCC becomes "rawcc".