fixes on enter-chroot script
$UID is not POSIX. to get the uid for the
current user, `id -u` must be called.
note that although $USER and $LOGNAME are
defined by POSIX, they don't always return
the current user, but the logged in one.
that means if you su to root, then $USER
might still be the logged in user.
`id -u` returns the uid for the current user.